refactor(tools): Migrate `VerifyPackageCurationsTask` to a Kotlin script
This is a proof-of-concept migration to show how verification would work
with stand-alone Kotlin scripts. Ultimately, the goal is to migrate all
tools to stand-alone Kotlin scripts in order to get rid of the Gradle
project that just serves as the infrastructure to run tools tasks.
