+++ tools/curations/build.gradle.kts
@@ -4,7 +4,6 @@ import org.ossreviewtoolkit.tools.curations.GenerateAzureSdkForNetCurationsTask
 import org.ossreviewtoolkit.tools.curations.GenerateDotNetRuntimeCurationsTask
 import org.ossreviewtoolkit.tools.curations.GenerateLicenseClassificationsTask
 import org.ossreviewtoolkit.tools.curations.VerifyPackageConfigurationsTask
-import org.ossreviewtoolkit.tools.curations.VerifyPackageCurationsTask
 
 tasks {
     register<GenerateLicenseClassificationsTask>("generateLicenseClassifications")
@@ -14,7 +13,6 @@ tasks {
     register<GenerateDotNetRuntimeCurationsTask>("generateDotNetRuntimeCurations")
 
     register<VerifyPackageConfigurationsTask>("verifyPackageConfigurations")
-    register<VerifyPackageCurationsTask>("verifyPackageCurations")
 
     register("verify") {
         group = "verification"
